Backend/requirements.txt
vertexai
pillow
flask
flask_cors
google-cloud-storage